move cbs xtpl fns - DO NOT REVIEW
